Centralized all installation specifics to a single module.
